+++ __test__/unique.spec.js
@@ -2,32 +2,32 @@ const { assert } = require('chai');
 const pure = require('../index');
 
 describe('unique.js', () => {
-    describe('unique()', () => {
+    describe('exec()', () => {
         it('is able to call a function with no arguments and return a result', () => {
-            const result = pure.unique(pure.internet.email);
+            const result = pure.unique.exec(pure.internet.email);
             assert.equal(typeof result, 'string');
         });
 
         it('is able to call a function with arguments and return a result', () => {
             // third argument is provider, or domain for email
-            const result = pure.unique(pure.internet.email, [{ firstName: 'a', lastName: 'b', provider: 'c'}]);
+            const result = pure.unique.exec(pure.internet.email, [{ firstName: 'a', lastName: 'b', provider: 'c' }]);
             assert.ok(result.match(/[@]c/));
         });
 
         it('is able to call same function with arguments and return a result', () => {
             // third argument is provider, or domain for email
-            const result = pure.unique(pure.internet.email, [{ firstName: 'a', lastName: 'b', provider: 'c'}]);
+            const result = pure.unique.exec(pure.internet.email, [{ firstName: 'a', lastName: 'b', provider: 'c' }]);
             assert.ok(result.match(/[@]c/));
         });
 
         it('is able to exclude results as array', () => {
-            const result = pure.unique(pure.internet.protocol, [], { exclude: ['https'] });
+            const result = pure.unique.exec(pure.internet.protocol, [], { exclude: ['https'] });
             assert.equal(result, 'http');
         });
 
         it('is able to limit unique call by maxTime in ms', () => {
             try {
-                pure.unique(pure.internet.protocol, [], {
+                pure.unique.exec(pure.internet.protocol, [], {
                     maxTime: 1,
                     maxRetries: 9999,
                     exclude: [
@@ -42,7 +42,7 @@ describe('unique.js', () => {
 
         it('is able to limit unique call by maxRetries', () => {
             try {
-                pure.unique(pure.internet.protocol, [], {
+                pure.unique.exec(pure.internet.protocol, [], {
                     maxTime: 5000,
                     maxRetries: 5,
                     exclude: [
@@ -57,8 +57,39 @@ describe('unique.js', () => {
 
         it('is able to call last function with arguments and return a result', () => {
             // third argument is provider, or domain for email
-            const result = pure.unique(pure.internet.email, [{ firstName: 'a', lastName: 'b', provider: 'c'}]);
+            const result = pure.unique.exec(pure.internet.email, [{ firstName: 'a', lastName: 'b', provider: 'c' }]);
             assert.ok(result.match(/[@]c/));
+        });
+    });
+
+    describe('clear()', () => {
+        it('is able to clear the found items at global scope', () => {
+            pure.unique.clear();
+            pure.unique.exec(pure.internet.protocol, [], {
+                exclude: ['https'],
+            });
+            pure.unique.clear();
+
+            const result = pure.unique.exec(pure.internet.protocol, [], {
+                exclude: ['https'],
+            });
+
+            assert.equal(result, 'http');
+        });
+
+        it('is able to clear the found items at functional scope', () => {
+            pure.unique.exec(pure.internet.protocol, [], {
+                exclude: ['https'],
+                scope: 'pureInternetProtocol',
+            });
+            pure.unique.clear('pureInternetProtocol');
+
+            const result = pure.unique.exec(pure.internet.protocol, [], {
+                exclude: ['https'],
+                scope: 'pureInternetProtocol',
+            });
+
+            assert.equal(result, 'http');
         });
     });
 });