+++ src/modules/unique.js
@@ -1,4 +1,4 @@
-const uniqueExec = require('../vendor/unique');
+const unique = require('../vendor/unique');
 
 /**
  *
@@ -7,7 +7,7 @@ const uniqueExec = require('../vendor/unique');
 class Unique {
     constructor() {
         /**
-         * unique
+         * exec
          *
          * @description Generate unique entries passing specific method, normally used inside loops
          * @param {Function} method Method that will be executed to generate data
@@ -19,12 +19,14 @@ class Unique {
          * @param {Array} [opts.exclude= empty] Global exclude list of results
          * @param {Function} [opts.compare] Uniqueness compare function, default behavior is to check
          *  value as key against object hash
-         * @method pure.unique
+         * @param {String} [opts.scope = null] Define what scope unique will be added, if nothing
+         * is passed then it will consider global scope
+         * @method pure.unique.exec
          * @example
          * let arr = [];
          *
          * for (let index = 0; index < 5; index++) {
-         *     arr.push(pure.unique(pure.internet.email))
+         *     arr.push(pure.unique.exec(pure.internet.email))
          * }
          *
          * console.log(arr);
@@ -37,7 +39,7 @@ class Unique {
          * //]
          *
          */
-        this.unique = (method, args, opts = {}) => {
+        this.exec = (method, args, opts = {}) => {
             const options = opts;
             options.startTime = new Date().getTime();
 
@@ -47,9 +49,45 @@ class Unique {
             if (typeof options.maxRetries !== 'number') {
                 options.maxRetries = 50;
             }
+            if (typeof options.scope !== 'string') {
+                options.scope = null;
+            }
 
             options.currentIterations = 0;
-            return uniqueExec.exec(method, args, options);
+            return unique.exec(method, args, options);
+        };
+
+        /**
+         * clear
+         *
+         * @description Clear previously generated entries to use inside another scope
+         * @param {String} [scope= null] Define what scope unique will cleared, if nothing
+         * is passed then it will consider global scope
+         * @method pure.unique.clear
+         * @example
+         * let result = pure.unique.exec(pure.internet.protocol, [], {
+         *      exclude: ['https'],
+         * });
+         *
+         * console.log(result);
+         * //outputs: http;
+         *
+         * pure.unique.clear();
+         *
+         * let result2 = pure.unique.exec(pure.internet.protocol, [], {
+         *      exclude: ['https'],
+         * });
+         *
+         * console.log(result2);
+         * //outputs: http;
+         *
+         */
+        this.clear = (scope) => {
+            if (typeof scope !== 'string') {
+                return unique.clear(null);
+            }
+
+            return unique.clear(scope);
         };
     }
 }